+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Fixed
+- Fixed an error that could occur when running `version` before `init` [#270](https://github.com/craftcms/nitro/issues/270).
 - Fixed an error that could occur when Nitro is unable to detect the database engine from a backup [#269](https://github.com/craftcms/nitro/issues/269).
 
 ## 2.0.3 - 2021-03-03